Remove `.latest` suffix from tag names
This suffix is pretty nonstandard, and looks ugly.  Typically `latest`
is used as the sole value of a tag given that the "type" of an image is
already covered in the "repo" part of the name (i.e.
`KhronosGroup/$DOCKERFILE:latest`).  `.$VERSION` remains as that is what
is already actively used.

Over time we can hopefully switch to the
`KhronosGroup/$DOCKERFILE:latest` + `KhronosGroup/$DOCKERFILE:$VERSION`
format across all images.