android-sdk: No longer install removed `patcher;v4`
The CI is failing on the `openxr-android` image because `patcher;v4`
does not exist anymore.  While it is unclear why this is gone, if there's
a replacement, or if this will even succeed, let's at least try to build
the image and see if it has any impact on our runtime.