Update asciidoctor-spec image to ruby:3.3.3
This greatly reduces the number of CVEs reported by Dockerhub and also
updates the toolchain to current versions of asciidoctor (which
apparently gives some better error messages now than it used to).

Because of the underlying Debian updates, Python packages can no longer
be installed at system level without a very ugly and brute-force
override. Instead this image sets up a Python virtual environment
containing all the packages, with just enough of an OS-level Python to
install venv itself. If you are following the image instructions to do
your own non-Docker install of the toolchain, you will need to setup a
venv and then use the 'activate' script or otherwise setup your
VIRTUAL_ENV and PATH to use that version. Scripts relying on Python
packages installed by this image must invoke the python3 in the venv,
not /usr/bin/python3.